+++ SampleCode/StdDriver/PDMA_TimerTrigger_SRAM2GPIO/main.c
@@ -153,9 +153,9 @@ int main()
     /* Set source address is au8SrcArray, Source increment size is 8 bits(one byte), destination address is PA->DOUT (no increment)*/
     PDMA_SetTransferAddr(PDMA, PDMAchannel, (uint32_t)g_au8SrcArray, PDMA_SAR_INC, (uint32_t) &(PA->DOUT), PDMA_DAR_FIX);
     /* Request source is timer 0 */
-    PDMA_SetTransferMode(PDMA, PDMAchannel, PDMA_TMR0, FALSE, NULL);
+    PDMA_SetTransferMode(PDMA, PDMAchannel, PDMA_TMR0, FALSE,(uint32_t) NULL);
     /* Transfer type is burst transfer and burst size is 4 */
-    PDMA_SetBurstType(PDMA, PDMAchannel, PDMA_REQ_SINGLE, NULL);
+    PDMA_SetBurstType(PDMA, PDMAchannel, PDMA_REQ_SINGLE,(uint32_t) NULL);
 
     /* Enable interrupt */
     PDMA_EnableInt(PDMA, PDMAchannel, PDMA_INT_TRANS_DONE);
